requirements_dev.txt
SQLAlchemy==1.4.41
sqlalchemy-file==0.1.4
SQLAlchemy-Utils==0.38.3
shortuuid==1.0.9
fasteners==0.18
PyYAML==6.0

black==22.8.0
ipdb==0.13.9
pipdeptree==2.3.1
pytest==7.1.3
tox==3.26.0

